src/packages/boan-computer-use/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
playwright==1.47.0
Pillow==10.4.0
httpx==0.27.2
